refactor(example): reorganised code
