+++ example/src/utils/hooks.tsx
@@ -1,0 +1,14 @@
+import { useLayoutEffect, useState } from "react";
+
+export const useWindowSize = () => {
+  const [width, setWidth] = useState(0);
+  useLayoutEffect(() => {
+    const updateSize = () => {
+      setWidth(window.innerWidth);
+    };
+    window.addEventListener("resize", updateSize);
+    updateSize();
+    return () => window.removeEventListener("resize", updateSize);
+  }, []);
+  return width;
+};
